Update indices for the nested CH and NA grids
Change the I, J indices for the nested CH and NA domains so that they
are consistent with the nested grid domains used for GEOS-5. We had
made the domains smaller for GEOS-FP because of the higher native
resolution (0.25x0.3125), but we can restore the domains to the full
extent of the China/SE Asia and North America regions since MERRA2 has
a native resolution of 0.5x0.6125.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>